+++ backend/tests/Functional/Controller/App/GroupControllerTest.php
@@ -101,7 +101,8 @@ class GroupControllerTest extends WebTestCase
 
         $this->assertSame([
             ['id' => $this->group1Id, 'name' => 'g1', 'description' => null,
-                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                'isDefault' => false, 'isAutoManaged' => null],
         ], $body1);
     }
 
@@ -140,7 +141,8 @@ class GroupControllerTest extends WebTestCase
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertSame([
             ['id' => $this->group0Id, 'name' => 'g0', 'description' => null,
-                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                'isDefault' => false, 'isAutoManaged' => null],
         ], $this->parseJsonBody($response));
     }
 
@@ -184,7 +186,8 @@ class GroupControllerTest extends WebTestCase
             'character' => ['id' => 123, 'name' => 'C1', 'corporation' => null],
             'groups' => [
                 ['id' => $this->group1Id, 'name' => 'g1', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
             ],
             'deactivated' => 'no',
         ], [
@@ -195,9 +198,11 @@ class GroupControllerTest extends WebTestCase
             ]],
             'groups' => [
                 ['id' => $this->group0Id, 'name' => 'g0', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
                 ['id' => $this->group1Id, 'name' => 'g1', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
             ],
             'deactivated' => 'no',
         ]];
@@ -231,7 +236,8 @@ class GroupControllerTest extends WebTestCase
             'character' => ['id' => 123, 'name' => 'C1', 'corporation' => null],
             'groups' => [
                 ['id' => $this->group1Id, 'name' => 'g1', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
             ],
             'deactivated' => 'no',
         ], [
@@ -277,7 +283,8 @@ class GroupControllerTest extends WebTestCase
             'character' => ['id' => 123, 'name' => 'C1', 'corporation' => null],
             'groups' => [
                 ['id' => $this->group1Id, 'name' => 'g1', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
             ],
             'deactivated' => 'no',
         ], [
@@ -288,9 +295,11 @@ class GroupControllerTest extends WebTestCase
             ]],
             'groups' => [
                 ['id' => $this->group0Id, 'name' => 'g0', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
                 ['id' => $this->group1Id, 'name' => 'g1', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
             ],
             'deactivated' => 'soon',
         ]];
@@ -354,9 +363,11 @@ class GroupControllerTest extends WebTestCase
 
         $this->assertSame([
             ['id' => $this->group1Id, 'name' => 'g1', 'description' => null,
-                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                'isDefault' => false, 'isAutoManaged' => null],
             ['id' => $this->group4Id, 'name' => 'g4', 'description' => null,
-                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                'isDefault' => false, 'isAutoManaged' => null],
         ], $this->parseJsonBody($response));
     }
 
@@ -399,16 +410,20 @@ class GroupControllerTest extends WebTestCase
         $expected = [[
             'id' => 500, 'name' => 'five', 'ticker' => '-5-', 'groups' => [
                 ['id' => $this->group1Id, 'name' => 'g1', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
                 ['id' => $this->group4Id, 'name' => 'g4', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
             ],
         ], [
             'id' => 501, 'name' => 'f1', 'ticker' => '-51-', 'groups' => [
                 ['id' => $this->group0Id, 'name' => 'g0', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
                 ['id' => $this->group1Id, 'name' => 'g1', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
             ],
         ]];
         $this->assertSame($expected, $body);
@@ -471,9 +486,11 @@ class GroupControllerTest extends WebTestCase
 
         $this->assertSame([
             ['id' => $this->group4Id, 'name' => 'g4', 'description' => null,
-                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                'isDefault' => false, 'isAutoManaged' => null],
             ['id' => $this->group5Id, 'name' => 'g5', 'description' => null,
-                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                'isDefault' => false, 'isAutoManaged' => null],
         ], $this->parseJsonBody($response));
     }
 
@@ -516,16 +533,20 @@ class GroupControllerTest extends WebTestCase
         $expected = [[
             'id' => 100, 'name' => 'one', 'ticker' => '-1-', 'groups' => [
                 ['id' => $this->group4Id, 'name' => 'g4', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
                 ['id' => $this->group5Id, 'name' => 'g5', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
             ],
         ], [
             'id' => 101, 'name' => 'o1', 'ticker' => '-11-', 'groups' => [
                 ['id' => $this->group0Id, 'name' => 'g0', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
                 ['id' => $this->group4Id, 'name' => 'g4', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null],
             ],
         ]];
         $this->assertSame($expected, $body);
@@ -563,7 +584,8 @@ class GroupControllerTest extends WebTestCase
         $body = $this->parseJsonBody($response);
         $this->assertSame([
             ['id' => $this->group1Id, 'name' => 'g1', 'description' => null,
-                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                'isDefault' => false, 'isAutoManaged' => null],
         ], $body);
     }
 
@@ -588,11 +610,14 @@ class GroupControllerTest extends WebTestCase
         $body = $this->parseJsonBody($response);
         $this->assertSame([
             ['id' => $this->group1Id, 'name' => 'g1', 'description' => null,
-                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                'isDefault' => false, 'isAutoManaged' => null],
             ['id' => $this->group4Id, 'name' => 'g4', 'description' => null,
-                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                'isDefault' => false, 'isAutoManaged' => null],
             ['id' => $this->group5Id, 'name' => 'g5', 'description' => null,
-                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false],
+                'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                'isDefault' => false, 'isAutoManaged' => null],
         ], $body);
     }