+++ backend/tests/Functional/Controller/User/AppControllerTest.php
@@ -211,7 +211,8 @@ class AppControllerTest extends WebTestCase
         $this->assertEquals(200, $response2->getStatusCode());
 
         $expectedGroup = ['id' => $this->gid, 'name' => 'group-one', 'description' => null,
-            'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false];
+            'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+            'isDefault' => false, 'isAutoManaged' => null];
         $this->assertSame(
             [
                 'id' => $this->aid,
@@ -443,7 +444,8 @@ class AppControllerTest extends WebTestCase
                 'id' => $this->aid,
                 'name' => 'app one',
                 'groups' => [['id' => $this->gid, 'name' => 'group-one', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false]],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null]],
                 'roles' => [Role::APP],
                 'eveLogins' => [[
                     'id' => $this->eveLoginId1,
@@ -470,7 +472,8 @@ class AppControllerTest extends WebTestCase
                 'id' => $this->aid,
                 'name' => 'app one',
                 'groups' => [['id' => $this->gid, 'name' => 'group-one', 'description' => null,
-                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false, 'isDefault' => false]],
+                    'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false,
+                    'isDefault' => false, 'isAutoManaged' => null]],
                 'roles' => [Role::APP],
                 'eveLogins' => [[
                     'id' => $this->eveLoginId1,